pyproject.toml
[project]
name = "xdist-custom-scheduler"
version = "0.1.0"
description = ""
authors = [
    {name = "jangil",email = "jangil.oh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-xdist (>=3.6.1,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
